I made a more proper draw() abstraction for DGImage.  It's still not totally what Tom wants, but if nothing else, it makes the current controls' draw methods a little saner (and less redundant).
I also moved over to using signed long rather than unsigned long as
the returns for DGImage's getHeight and getWidth methods and also as
the port height argument for the various DGRect -> CGRect and draw
methods.  This fixed a few bugs that it turns out were happening
during some arithmatic where the entire operations were quietly being
cast to unsigned long rather than signed long, and wraparound was
occurring (for example, that's what was causing the weird clipping in
AudioUnitHosting when you opened an AU that had a UI view that was
taller than the current size of AUH's window).

I also trid experimenting with transforming the CoreGraphics drawing
matrix so that I could hopefully use non-upside-down coordinates.  I
was able to this and successfully position things with non-upside-down
coordinates, but the big drawback was that everything that got drawn
(images and text) was drawn upside down.  Hmmm...

+++ dfxgui/dfxguieditor.h
@@ -46,7 +46,7 @@ public:
 	void do_idle();
 	virtual void idle() { }
 
-	virtual void DrawBackground(CGContextRef inContext, UInt32 inPortHeight);
+	virtual void DrawBackground(CGContextRef inContext, long inPortHeight);
 
 #if MAC
 	virtual bool HandleMouseEvent(EventRef inEvent);
